requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# STYLiSH/app/__init__.py: 1
# STYLiSH/app/main/__init__.py: 1
# STYLiSH/app/main/controller/admin_controller.py: 3
# STYLiSH/app/main/controller/dashboard_controller.py: 2
# STYLiSH/app/main/controller/fb_login_controller.py: 1
# STYLiSH/app/main/controller/page_controller.py: 1
# STYLiSH/app/main/controller/product_controller.py: 2
# STYLiSH/app/main/controller/recom_controller.py: 3
# STYLiSH/app/main/controller/user_controller.py: 2
# STYLiSH/app/main/service/admin.py: 3
# STYLiSH/app/main/service/dashboard.py: 3
# STYLiSH/app/main/service/product.py: 3
# STYLiSH/app/main/service/recom.py: 3
# STYLiSH/app/main/service/user.py: 3
# STYLiSH/app/test/test_config.py: 4
Flask == 1.1.1

# STYLiSH/app/main/__init__.py: 3
# STYLiSH/app/main/controller/user_controller.py: 6
Flask_Bcrypt == 0.7.1

# STYLiSH/app/__init__.py: 2
# STYLiSH/app/main/controller/admin_controller.py: 5
# STYLiSH/app/main/controller/dashboard_controller.py: 4
# STYLiSH/app/main/controller/fb_login_controller.py: 2
# STYLiSH/app/main/controller/page_controller.py: 2
# STYLiSH/app/main/controller/product_controller.py: 4
# STYLiSH/app/main/controller/recom_controller.py: 5
# STYLiSH/app/main/controller/user_controller.py: 4
Flask_RESTful == 0.3.8

# STYLiSH/manage.py: 5
Flask_Script == 2.0.6

# STYLiSH/app/test/test_config.py: 5
Flask_Testing == 0.8.1

# STYLiSH/app/main/service/user.py: 5
PyJWT == 2.0.1

# STYLiSH/app/main/controller/admin_controller.py: 6
# STYLiSH/app/main/controller/dashboard_controller.py: 5
# STYLiSH/app/main/controller/product_controller.py: 5
# STYLiSH/app/main/controller/recom_controller.py: 6
# STYLiSH/app/main/controller/user_controller.py: 5
# STYLiSH/app/main/service/admin.py: 1
# STYLiSH/app/main/service/dashboard.py: 1
# STYLiSH/app/main/service/product.py: 1
# STYLiSH/app/main/service/recom.py: 1
# STYLiSH/app/main/service/user.py: 1,4
PyMySQL == 1.0.2

# STYLiSH/app/main/controller/admin_controller.py: 1
# STYLiSH/app/main/controller/dashboard_controller.py: 1
# STYLiSH/app/main/controller/product_controller.py: 1
# STYLiSH/app/main/controller/recom_controller.py: 1
# STYLiSH/app/main/controller/user_controller.py: 1
Werkzeug == 1.0.0

# STYLiSH/app/main/controller/admin_controller.py: 2
# STYLiSH/app/main/controller/recom_controller.py: 2
boto3 == 1.17.46

# STYLiSH/app/main/controller/fb_login_controller.py: 4
# STYLiSH/app/main/service/admin.py: 2
# STYLiSH/app/main/service/dashboard.py: 2
# STYLiSH/app/main/service/product.py: 2
# STYLiSH/app/main/service/recom.py: 2
# STYLiSH/app/main/service/user.py: 2
python_dotenv == 0.15.0

# STYLiSH/app/main/controller/dashboard_controller.py: 10
# STYLiSH/app/main/controller/user_controller.py: 12
# STYLiSH/app/main/service/admin.py: 5
# STYLiSH/app/main/service/dashboard.py: 5
# STYLiSH/app/main/service/product.py: 5
# STYLiSH/app/main/service/recom.py: 5
# STYLiSH/app/main/service/user.py: 7
pytz == 2019.3

# STYLiSH/app/__init__.py: 3
# STYLiSH/app/main/controller/recom_controller.py: 7
redis == 3.5.3

# STYLiSH/app/main/controller/user_controller.py: 8
requests == 2.22.0

# STYLiSH/app/main/controller/fb_login_controller.py: 5,6
requests_oauthlib == 1.3.0
